# speed up `String::push` and `String::insert`