@lincot 2024-05-13
A codegen check for the absence of `memcpy` would be nice, since the original `String::push` has one.