@lincot 2024-10-06
The proposed implementation uses `get_unchecked_mut`, which cannot be used in `char::encode_utf8`, which is now `const`. Also, `get_unchecked_mut` actually has a cost when running with debug assertions. So I am reverting to using pointers. The codegen of `String::push` seems to be unchanged: [godbolt](https://rust.godbolt.org/z/cdTv1K1rE).