@lincot 2025-04-05
The benchmarks for `String::push` in `library/alloctests/benches/string.rs` seem to lack `black_box`.

<details>

<summary>Original benchmark results</summary>

before the patch:

```
string::bench_push_char_one_byte       10464.93ns/iter +/- 104.38
string::bench_push_char_two_bytes       8793.83ns/iter  +/- 33.15
string::bench_insert_char_long            40.98ns/iter   +/- 0.82
string::bench_insert_char_short           40.19ns/iter   +/- 0.35
```

after the patch:

```
string::bench_push_char_one_byte       10247.51ns/iter +/- 18.39
string::bench_push_char_two_bytes       9575.89ns/iter +/- 40.07
string::bench_insert_char_long            34.36ns/iter  +/- 2.17
string::bench_insert_char_short           34.17ns/iter  +/- 0.57
```

</details>

If we make the design of the `String::push` benchmarks similar to the `String::insert` benchmarks, the results begin to convey the difference, especially when pushing a two-byte character.

<details>

<summary>The improved benchmark code</summary>

```rs
#[bench]
fn bench_push_char_two_bytes(b: &mut Bencher) {
    b.bytes = REPETITIONS * 2;
    b.iter(|| {
        let mut r = String::new();
        for _ in 0..REPETITIONS {
            black_box(&mut r).push(black_box('â'));
        }
        r
    });
}
```

</details>

before the patch:

```
string::bench_push_char_one_byte       12670.38ns/iter +/- 151.43
string::bench_push_char_two_bytes      49854.63ns/iter +/- 367.59
string::bench_insert_char_long            40.36ns/iter   +/- 0.57
string::bench_insert_char_short           39.40ns/iter   +/- 0.63
```

after the patch:

```
string::bench_push_char_one_byte       12872.57ns/iter +/- 122.83
string::bench_push_char_two_bytes      14988.39ns/iter +/- 120.85
string::bench_insert_char_long            35.46ns/iter   +/- 0.50
string::bench_insert_char_short           35.04ns/iter   +/- 0.48
```

I can make an additional commit that improves the benchmark.