@lincot 2025-04-06
In the message above you can observe a regression in `string::bench_push_char_one_byte`. I decided to run the benchmarks 10 more times per implementation:

without the patch: 12745.60, 12654.10, 12654.79, 12652.53, 12669.58, 12751.50, 12656.31, 12654.54, 12652.73, 12650.33 (average 12674.2 ns)
with the patch: 12655.30, 12657.40, 12829.48, 12651.96, 12653.84, 12653.74, 12655.01, 12660.19, 12664.05, 12664.41 (average 12674.5 ns)

Both give the same average so we can say that the performance doesn't change for a single-byte push.

I also tried to add a `likely` for the single-byte length in the `len_utf8` function, the average time did not change.

For a two-byte push, though, the average time goes from 49850ns to 14785ns, giving a 3.37x speedup.